Permissions
Here, the extension declares the permissions it requires. In this case, it requests access to the active tab and the ability to use declarative content. These permissions allow the extension to interact with the currently active tab in the browser and to use declarative rules to modify the content of web pages.